Add new file for missions API and create missions rducer
